Deployment/Prediction Service/requirements.txt
PySide6
Vispy
Scipy
pythonnet
numpy>=1.21.0
paho.mqtt.client 
time
socket
pandas>=1.3.0
csv
json
matplotlib>=3.4.0
libemg
scipy>=1.7.0
pywavelets>=1.1.1
plotly>=5.0.0
seaborn>=0.11.0
scikit-learn>=0.24.0
tensorflow>=2.5.0
kafka-python>=2.0.2
influxdb-client>=1.23.0
flask>=2.0.0
psutil>=5.8.0
fastapi
uvicorn
